// generated-by-analysis/Pilot9/graph-inputs/harm-quantity-with-likert-severity-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/harm-quantity-with-likert-severity.ts.
// spell-checker: disable

	export const LostPhotos = {
	"labels": [
		"None",
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			1,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			2,
			1
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			1,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			2,
			0
		],
		"No loss": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			10,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			20,
			10
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			10,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			20,
			0
		],
		"No loss": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		]
	}
};
	export const LostEmails = {
	"labels": [
		"None",
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			1,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			1,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0,
			1
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0,
			1
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"No loss": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			10,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			10,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0,
			10
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0,
			10
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"No loss": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		]
	}
};